requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
yfinance>=0.2.0
pytz>=2024.1
discord.py>=2.3.0
aiohttp>=3.9.0
apscheduler>=3.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
